+++ features/poolFinder/types.ts
@@ -1,0 +1,6 @@
+export interface OraclessPoolResult {
+  collateralAddress: string
+  collateralToken: string
+  quoteAddress: string
+  quoteToken: string
+}
